+++ src/simulation/Collision.cpp
@@ -40,7 +40,7 @@ void Collision::generateFragments() {
     double fragmentCount = 0.1 * std::pow(mass, 0.75) * std::pow(_minimalCharacteristicLength, -1.71);
 
     const std::string debrisName{sat1.getName() + " & " + sat2.getName() + "-Collision-Fragment"};
-    this->createFragments(fragmentCount, debrisName);
+    this->createFragments(static_cast<size_t>(fragmentCount), debrisName);
 }
 
 void Collision::characteristicLengthDistribution() {
